refactor: domain 코드 리팩토링(허용어, 금칙어, 로그인 관련 제외)
