+++ src/main/java/com/ureca/uplait/domain/admin/controller/AdminReviewController.java
@@ -4,7 +4,7 @@ import static org.springframework.data.domain.Sort.Direction;
 
 import com.ureca.uplait.domain.admin.dto.response.AdminReviewDeleteResponse;
 import com.ureca.uplait.domain.admin.service.AdminReviewService;
-import com.ureca.uplait.domain.review.dto.response.AdminReviewResponse;
+import com.ureca.uplait.domain.admin.dto.response.AdminReviewResponse;
 import com.ureca.uplait.global.response.CommonResponse;
 import com.ureca.uplait.global.response.ResultCode;
 import io.swagger.v3.oas.annotations.Operation;
