+++ src/main/java/com/ureca/uplait/domain/plan/dto/request/PlanCompareRequest.java
@@ -9,8 +9,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Schema(description = "요금제 비교 요청 DTO")
 public class PlanCompareRequest {
-
     @Schema(description = "비교할 요금제의 ID 리스트", example = "[1, 2, 3]")
     private List<Long> planIds;
-
 }
