+++ src/main/java/com/ureca/uplait/domain/plan/dto/response/PlanDetailResponse.java
@@ -9,7 +9,6 @@ import lombok.Getter;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @Schema(description = "요금제 상세 조회 결과")
 public abstract class PlanDetailResponse {
-
     @Schema(description = "요금제 id", example = "1")
     protected Long planId;
