+++ src/main/java/com/ureca/uplait/domain/plan/dto/response/PlanListResponse.java
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Schema(description = "요금제 목록 조회 결과 (이름, 가격만 포함)")
 public class PlanListResponse {
-
     @Schema(description = "요금제 id", example = "1")
     private Long planId;
