+++ src/main/java/com/ureca/uplait/domain/plan/repository/IPTVPlanRepository.java
@@ -1,11 +1,10 @@
 package com.ureca.uplait.domain.plan.repository;
 
 import com.ureca.uplait.domain.plan.entity.IPTVPlan;
-import com.ureca.uplait.domain.plan.entity.MobilePlan;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPTVPlanRepository extends JpaRepository<MobilePlan, Long> {
+public interface IPTVPlanRepository extends JpaRepository<IPTVPlan, Long> {
 
     List<IPTVPlan> findAllByOrderByIdDesc();
 }
