+++ src/main/java/com/ureca/uplait/domain/user/controller/BookmarkController.java
@@ -20,14 +20,6 @@ public class BookmarkController {
 
     private final BookmarkService bookmarkService;
 
-    /**
-     *  즐겨찾기 조회
-     *
-     * @param user 사용자 정보
-     * @param planType 요금제 종류
-     * @param size 페이지 크기
-     * @param lastBookmarkId 마지막 즐겨찾기 id
-     */
     @Operation(summary = "즐겨찾기 조회", description = "즐겨찾기 조회: 로그인 필요")
     @GetMapping
     public CommonResponse<BookmarkListResponse> getBookmarks(
@@ -42,12 +34,6 @@ public class BookmarkController {
         return CommonResponse.success(bookmarkService.getBookmarks(user, planType, size, lastBookmarkId));
     }
 
-    /**
-     *  즐겨찾기 추가
-     *
-     * @param user 사용자 정보
-     * @param planId 요금제 Id
-     */
     @Operation(summary = "즐겨찾기 추가", description = "즐겨찾기 추가: 로그인 필요")
     @PostMapping
     public CommonResponse<CreateBookmarkResponse> createBookmark(
@@ -58,12 +44,6 @@ public class BookmarkController {
         return CommonResponse.success(bookmarkService.createBookmark(user, planId));
     }
 
-    /**
-     *  즐겨찾기 삭제
-     *
-     * @param user 사용자 정보
-     * @param bookmarkId 삭제할 bookmark Id
-     */
     @Operation(summary = "즐겨찾기 삭제", description = "즐겨찾기 삭제: 로그인 필요")
     @DeleteMapping("/{bookmarkId}")
     public CommonResponse<DeleteBookmarkResponse> deleteBookmark(
