+++ src/main/java/com/ureca/uplait/domain/user/dto/request/ExtraInfoRequest.java
@@ -15,7 +15,6 @@ import lombok.Setter;
 @Setter
 @Schema(description="회원가입 시 추가정보 담는 DTO")
 public class ExtraInfoRequest {
-
 	@Schema(description = "전화번호", example = "010-1234-5678")
 	@Pattern(regexp = "^010-\\d{4}-\\d{4}$", message = "전화번호는 010-XXXX-XXXX 형식이어야 합니다.")
 	String phoneNumber;
@@ -35,5 +34,4 @@ public class ExtraInfoRequest {
 
 	@Schema(description = "광고수신동의여부", example = "true")
 	boolean adAgree;
-
 }
